Extract base midi methods to seperate class
